Move some effect calc functions out of player
